Enhance tree of cards
Create tree of cards with custom node separator. Separator
customized on Options page and used to split labels of cards
as a path in a tree. Example: label "a/b/c" with separator "/"
will produce subtree with nodes a, b, c, where b is a child of a and
c is a child of b.